fix: 64 bit warnings fix: other small fixes change: more detail example change: animation API to JTabBarController add: hide tabbar

+++ JJlibUI/Examples/TabBarControllerExampleVC.m
@@ -83,7 +83,7 @@
 
 #pragma mark - JTabBarControllerDelegate
 
-- (UIButton *)tabBarController:(JTabBarController *)tabBarController tabBarButtonForChildViewController:(UIViewController *)childViewController forIndex:(uint)index {
+- (UIButton *)tabBarController:(JTabBarController *)tabBarController tabBarButtonForChildViewController:(UIViewController *)childViewController forIndex:(NSInteger)index {
     
     // this will only execute if the childViewController do not provide a jTabBarButton
     return [[NSBundle mainBundle] loadNibNamed:@"AltMainButtonTemplate" owner:self options:nil][0];;
